`tree::walk`: Allow callback to return an int
`git_tree_walk` in libgit2 accepts a callback which can return an int.
The int tells the walker whether to recurse on a tree object or not.

This is useful behaviour if you don't want to walk an *entire* large
tree.

To upgrade existing code, just change existing callbacks to return `int`
instead of `void`, and add a `return 0;`