Require libgit2 1.4.0+. Remove data_buffer methods
This change is slightly backward incompatible. libgit2 1.4.0 made a
chagne to the `git_buf` structure and deprecated a bunch of associated
functions, to the effect that `git_buf` is no longer modifiable by
users. It is expected that `git_buf` should only be set by libgit2
itself as a return from various functions.

So this change to the `data_buffer` class removes all modifier methods.
A data_buffer is returned where libgit2 returns a `git_buf`, but we no
longer allow the caller to modify the buffer data or grow the buffer.

related: https://github.com/libgit2/libgit2/pull/6078

Perhaps we should change methods returning a `data_buffer` to return a
`std::string` instead?